verification email front, sequelize update et delete, front admin

+++ Front/groupomania-project/src/components/NavBar.vue
@@ -1,11 +1,15 @@
 <template>
   
-  <nav v-if="isLogged">
+  
+  <ChildComponent @name="getName"/>
+	<nav v-if="isLogged">
     <router-link to="/homepost">Forum</router-link> |
     <router-link to="/profile">Profile</router-link> |
-    <router-link to="/"><span @click="logOut()">Déconnexion</span></router-link>
-
+    <router-link to="/alluser" v-if="admAc">Abonné  | </router-link> 
+    <router-link to="/" @click="logOut()">Déconnexion</router-link>
+    
   </nav>
+  
    <nav v-else>
     <router-link to="/">Accueil</router-link> |
     <router-link to="/login">Connexion</router-link> |
@@ -18,33 +22,87 @@
 </template>
 
 <script>
+
+import { mapState } from 'vuex'
+
 export default {
   name: 'NavBar',
-  props: {
-    msg: String
+  data() {
+    return {
+      userData:'',
+    }
   },
+
+   props :{
+        outLog: Boolean,
+      },
+
+    mounted : function() {
+        if(localStorage.getItem('user') != null) {
+          let userData = JSON.parse(localStorage.getItem("user"));
+          this.userData= userData;
+
+          }
+    },
+      
    computed: {
+   admAc (){
+        if(localStorage.getItem('user') != null) {
+           let userAdd = JSON.parse(localStorage.getItem("user"));
+          //console.log(userAdd.nom, " nom");
+          if(userAdd.isAdmin == 49) {
+            return true;
+          }else{
+            return false
+          }
+      
+        }  else{
+            return false
+          }
+    }, 
+ 
     isLogged (){
         if(localStorage.getItem('user') != null) {
+          
           return true;
-        } else {
+        } else if(localStorage.getItem('user') == 'undefined'){
+          localStorage.clear();
           return false;
+          
         }
-      
-    },
+        else {
+          return false;
+        }  
+    }//logged
+   ,
+   ...mapState({
+      user: 'userInfos',
+    })
+    ,
     methods: {
+
       async logOut (){
         const self=this;
         this.$store.dispatch('logOut', {
           
       }).then( () => {
-        localStorage.clear(),
+        //this.$emit('isConnected', false),
         self.$router.push('/');
       }, (error) =>{
         console.log(error);
       }
       )
     } //logOut 
+    ,
+    getName(value) {
+      console.log(value); // Raja Tamil
+   }
+    
+     /* async logOut (){
+        window.localStorage.clear();
+        this.$emit('isConnected', false)
+        self.$router.go('/');
+    } //logOut */
   }//methode
 
   }
